Change angles from int to float

+++ source/ray/ray.h
@@ -6,7 +6,7 @@
 /*   By: pstengl <pstengl@student.42wolfsburg.      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/24 11:22:49 by pstengl           #+#    #+#             */
-/*   Updated: 2022/04/24 11:45:02 by pstengl          ###   ########.fr       */
+/*   Updated: 2022/05/02 15:47:53 by pstengl          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,6 @@ typedef struct s_ray {
 	char	side;
 }	t_ray;
 
-t_ray	ray_cast(t_map map, t_point pos, int angle);
+t_ray	ray_cast(t_map map, t_point pos, double angle);
 
 #endif
